requirements.txt
Flask
flask-cors
python-dotenv
requests
PyPDF2
python-docx